doc: 파일 분리.
